moved imagepanel into its own file to follow SRP
